Codigo/APIv3/requirements.txt
torch
ultralytics
fastapi
uvicorn
pillow
python-multipart